@@ -26,85 +26,86 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.2.3",
-    "@angular/cdk": "^13.2.3",
-    "@angular/common": "~13.2.3",
-    "@angular/compiler": "~13.2.3",
-    "@angular/core": "~13.2.3",
-    "@angular/forms": "~13.2.3",
-    "@angular/material": "^13.2.3",
-    "@angular/platform-browser": "~13.2.3",
-    "@angular/platform-browser-dynamic": "~13.2.3",
-    "@angular/router": "~13.2.3",
-    "@danielmoncada/angular-datetime-picker": "^13.1.1",
-    "@muni-kypo-crp/adaptive-transition-visualization": "~13.0.0",
-    "@muni-kypo-crp/adaptive-visualization": "~13.0.1",
-    "@muni-kypo-crp/assessment-visualization": "~13.0.0",
-    "@muni-kypo-crp/command-visualizations": "~13.0.1",
-    "@muni-kypo-crp/d3-service": "13.0.0",
-    "@muni-kypo-crp/hurdling-visualization": "~13.1.0",
-    "@muni-kypo-crp/overview-visualization": "~13.0.2",
-    "@muni-kypo-crp/sandbox-api": "~13.0.0",
-    "@muni-kypo-crp/sandbox-model": "~13.0.0",
-    "@muni-kypo-crp/statistical-visualizations": "~13.0.1",
-    "@muni-kypo-crp/topology-graph": "~13.0.2",
-    "@muni-kypo-crp/topology-model": "~13.0.0",
-    "@muni-kypo-crp/training-api": "~13.0.7",
-    "@muni-kypo-crp/training-model": "~13.0.8",
-    "@muni-kypo-crp/visualization-dashboard": "~13.0.1",
-    "@sentinel/auth": "~13.0.0",
-    "@sentinel/common": "~13.0.0",
-    "@sentinel/components": "~13.3.0",
-    "@sentinel/layout": "~13.0.2",
+    "@angular/animations": "^14.0.3",
+    "@angular/cdk": "^14.0.3",
+    "@angular/common": "^14.0.3",
+    "@angular/compiler": "^14.0.3",
+    "@angular/core": "^14.0.3",
+    "@angular/forms": "^14.0.3",
+    "@angular/material": "^14.0.3",
+    "@angular/platform-browser": "^14.0.3",
+    "@angular/platform-browser-dynamic": "^14.0.3",
+    "@angular/router": "^14.0.3",
+    "@danielmoncada/angular-datetime-picker": "^14.1.0",
+    "@muni-kypo-crp/adaptive-transition-visualization": "~14.0.0",
+    "@muni-kypo-crp/adaptive-visualization": "~14.0.0",
+    "@muni-kypo-crp/assessment-visualization": "~14.0.0",
+    "@muni-kypo-crp/command-visualizations": "~14.0.0",
+    "@muni-kypo-crp/d3-service": "14.0.0",
+    "@muni-kypo-crp/hurdling-visualization": "~14.0.0",
+    "@muni-kypo-crp/overview-visualization": "~14.0.0",
+    "@muni-kypo-crp/sandbox-api": "~14.0.0",
+    "@muni-kypo-crp/sandbox-model": "~14.0.0",
+    "@muni-kypo-crp/statistical-visualizations": "~14.0.0",
+    "@muni-kypo-crp/topology-graph": "~14.0.1",
+    "@muni-kypo-crp/topology-model": "~14.0.0",
+    "@muni-kypo-crp/training-api": "~14.0.0",
+    "@muni-kypo-crp/training-model": "~14.0.0",
+    "@muni-kypo-crp/visualization-dashboard": "~14.0.0",
+    "@sentinel/auth": "~14.0.0",
+    "@sentinel/common": "~14.0.1",
+    "@sentinel/components": "~14.0.3",
+    "@sentinel/layout": "~14.0.0",
     "angular-file": "^3.6.0",
     "angular-mgl-timeline": "^0.4.0",
     "angular-oauth2-oidc": "^13.0.1",
     "angular-oauth2-oidc-jwks": "^13.0.1",
-    "d3": "^7.1.1",
-    "d3-graphviz": "^2.6.1",
+    "d3": "^7.5.0",
+    "d3-graphviz": "^4.1.1",
     "d3-sankey": "^0.12.3",
     "file-saver": "^2.0.5",
-    "marked": "~1.1.0",
+    "marked": "~4.0.17",
     "moment-mini": "^2.24.0",
-    "ngx-indexed-db": "^9.4.2",
-    "ngx-markdown": "^12.0.0",
-    "rxjs": "~7.5.2",
-    "tslib": "^2.3.1",
+    "ngx-indexed-db": "^11.0.2",
+    "ngx-markdown": "^14.0.1",
+    "rxjs": "~7.5.5",
+    "tslib": "^2.4.0",
     "typescript-collections": "~1.3.3",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.11.6",
+    "caret-pos": "^2.0.0"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.1302.4",
-    "@angular-devkit/build-angular": "~13.2.4",
-    "@angular-eslint/builder": "~13.0.1",
-    "@angular-eslint/eslint-plugin": "^13.0.1",
-    "@angular-eslint/eslint-plugin-template": "^13.0.1",
-    "@angular-eslint/schematics": "^13.0.1",
-    "@angular-eslint/template-parser": "^13.0.1",
-    "@typescript-eslint/eslint-plugin": "^5.10.1",
-    "@typescript-eslint/parser": "^5.10.1",
-    "@angular/cli": "~13.2.4",
-    "@angular/compiler-cli": "~13.2.3",
-    "@angular/language-service": "~13.2.3",
-    "@nrwl/cypress": "~12.3.3",
-    "@nrwl/workspace": "~12.3.3",
-    "@types/jasmine": "~3.8.0",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^17.0.8",
-    "cypress": "^6.9.0",
-    "eslint": "^8.6.0",
-    "husky": "^4.2.5",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.3.3",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "moment": "^2.27.0",
-    "ng-packagr": "^13.2.0",
-    "prettier": "^2.5.1",
-    "ts-node": "~8.3.0",
-    "typescript": "~4.5.5"
+    "@angular-devkit/architect": "^0.1400.3",
+    "@angular-devkit/build-angular": "^14.0.3",
+    "@angular-eslint/builder": "~14.0.0",
+    "@angular-eslint/eslint-plugin": "^14.0.0",
+    "@angular-eslint/eslint-plugin-template": "^14.0.0",
+    "@angular-eslint/schematics": "^14.0.0",
+    "@angular-eslint/template-parser": "^14.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.30.0",
+    "@typescript-eslint/parser": "^5.30.0",
+    "@angular/cli": "^14.0.3",
+    "@angular/compiler-cli": "^14.0.3",
+    "@angular/language-service": "^14.0.3",
+    "@nrwl/cypress": "~14.3.6",
+    "@nrwl/workspace": "~14.3.6",
+    "@types/jasmine": "~4.0.3",
+    "@types/jasminewd2": "~2.0.10",
+    "@types/node": "^18.0.0",
+    "cypress": "^10.2.0",
+    "eslint": "^8.18.0",
+    "husky": "^8.0.1",
+    "jasmine-core": "~4.2.0",
+    "jasmine-spec-reporter": "~7.0.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.1.1",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "^2.0.0",
+    "moment": "^2.29.3",
+    "ng-packagr": "^14.0.2",
+    "prettier": "^2.7.1",
+    "ts-node": "~10.8.1",
+    "typescript": "~4.7.4"
   }
 }
