@@ -1,15 +1,15 @@
-import { FormControl, FormGroup } from '@angular/forms';
+import { UntypedFormControl, UntypedFormGroup } from '@angular/forms';
 import { SentinelValidators } from '@sentinel/common';
 
 /**
  * Form group controls for form in clone dialog component
  */
 export class CloneDialogFormGroup {
-  formGroup: FormGroup;
+  formGroup: UntypedFormGroup;
 
   constructor() {
-    this.formGroup = new FormGroup({
-      clonedDefinitionTitle: new FormControl('', [SentinelValidators.noWhitespace]),
+    this.formGroup = new UntypedFormGroup({
+      clonedDefinitionTitle: new UntypedFormControl('', [SentinelValidators.noWhitespace]),
     });
   }
 }
